backend/scraper-service/requirements.txt
fastapi
uvicorn
google-api-python-client
requests
python-dotenv
